% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{AABMI}
\alias{AABMI}
\title{This function provides estimates of Age-Adjusted BMI(AABMI) using  BMI, age}
\usage{
AABMI(BMI, age)
}
\arguments{
\item{BMI}{body mass index}

\item{age}{age}

\item{AABMI}{Age Adjusted BMI}
}
\description{
This function provides estimates of Age-Adjusted BMI(AABMI) using  BMI, age
}
