% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{ABSI}
\alias{ABSI}
\title{This function provides estimates of body shape index (ABSI) using W, H, WC}
\usage{
ABSI(WC, H, W)
}
\arguments{
\item{WC}{waist circumference}

\item{H}{height}

\item{W}{weight}

\item{ABSI}{body shape index}
}
\description{
This function provides estimates of body shape index (ABSI) using W, H, WC
}
