% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{ABW}
\alias{ABW}
\title{This function provides estimates of Body Weight (ABW) using  W,H,sex}
\usage{
ABW(W, H, sex)
}
\arguments{
\item{W}{weight}

\item{H}{height}

\item{sex}{sex}

\item{ABW}{Body Weight}
}
\description{
This function provides estimates of Body Weight (ABW) using  W,H,sex
}
