% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{AC}
\alias{AC}
\title{This function provides estimates of Atherogenic Coefficient (AC) using total_chol_mgdl, HDL_mgdl}
\usage{
AC(total_chol_mgdl, HDL_mgdl)
}
\arguments{
\item{total_chol_mgdl, }{total cholesterol (mgdl)}

\item{HDL_mgdl}{high density liproproten (mgdl)}

\item{AC}{Atherogenic Coefficient (AC)}
}
\description{
This function provides estimates of Atherogenic Coefficient (AC) using total_chol_mgdl, HDL_mgdl
}
