% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{ACI}
\alias{ACI}
\title{This function provides estimates of Abdominal Circumference Index (ACI) using  WC, HC, H}
\usage{
ACI(WC, HC, H)
}
\arguments{
\item{WC}{waist circumference}

\item{HC}{hip circumference}

\item{H}{height}

\item{ACI}{Abdominal Circumference Index}
}
\description{
This function provides estimates of Abdominal Circumference Index (ACI) using  WC, HC, H
}
