% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{AGR}
\alias{AGR}
\title{This function provides estimates of Android-to-Gynoid Fat Ratio (AGR) using  WC, HC}
\usage{
AGR(WC, HC)
}
\arguments{
\item{WC}{waist circumference}

\item{HC}{hip circumference}

\item{AGR}{Android-to-Gynoid Fat Ratio}
}
\description{
This function provides estimates of Android-to-Gynoid Fat Ratio (AGR) using  WC, HC
}
