% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{AIBW}
\alias{AIBW}
\title{This function provides estimates of Adjusted Ideal Body Weight (AIBW) using H,sex}
\usage{
AIBW(H, sex)
}
\arguments{
\item{H}{height}

\item{sex}{sex}

\item{AIBW}{Adjusted Ideal Body Weight}
}
\description{
This function provides estimates of Adjusted Ideal Body Weight (AIBW) using H,sex
}
