% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{AIP}
\alias{AIP}
\title{This function provides estimates of Atherogenic Index of Plasma (AIP) using TG_mgdl, HDL_mgdl}
\usage{
AIP(TG_mgdl, HDL_mgdl)
}
\arguments{
\item{TG_mgdl, }{triglycerides (mgdl)}

\item{HDL_mgdl}{high density liproprotein (mgdl)}

\item{AIP}{Atherogenic Index of Plasma}
}
\description{
This function provides estimates of Atherogenic Index of Plasma (AIP) using TG_mgdl, HDL_mgdl
}
