% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{AMA}
\alias{AMA}
\title{This function provides estimates of Arm Muscle Area (AMA) using midarm_circumference, triceps_skinfold_thickness}
\usage{
AMA(midarm_circumference, triceps_skinfold_thickness)
}
\arguments{
\item{midarm_circumference}{midarm_circumference}

\item{triceps_skinfold_thickness}{triceps_skinfold_thickness}

\item{AMA}{Arm Muscle Area}
}
\description{
This function provides estimates of Arm Muscle Area (AMA) using midarm_circumference, triceps_skinfold_thickness
}
