% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{AST}
\alias{AST}
\title{This function provides estimates of Abdominal Skinfold Thickness (AST) using triceps, subscapular, suprailiac}
\usage{
AST(triceps, subscapular, suprailiac)
}
\arguments{
\item{triceps}{triceps}

\item{subscapular}{subscapular}

\item{suprailiac}{suprailiac}

\item{AST}{Abdominal Skinfold Thickness}
}
\description{
This function provides estimates of Abdominal Skinfold Thickness (AST) using triceps, subscapular, suprailiac
}
