% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{AVI}
\alias{AVI}
\title{This function provides estimates of adiposity volume index (AVI) using HC and WC}
\usage{
AVI(WC, HC)
}
\arguments{
\item{WC}{waist circumference}

\item{HC}{hip circumference}

\item{AVI}{adiposity volume index}
}
\description{
This function provides estimates of adiposity volume index (AVI) using HC and WC
}
