% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{BAI}
\alias{BAI}
\title{This function provides estimates of body adiposity index (BAI) using H, HC}
\usage{
BAI(HC, H)
}
\arguments{
\item{HC}{hip circumference}

\item{H}{height}

\item{BAI}{body adiposity index}
}
\description{
This function provides estimates of body adiposity index (BAI) using H, HC
}
