% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{BCM}
\alias{BCM}
\title{This function provides estimates of  Body Cell Mass (BCM) using sex, H,W}
\usage{
BCM(sex, H, W)
}
\arguments{
\item{sex}{sex}

\item{H}{height}

\item{W}{weight}

\item{BCM}{Body Cell Mass}
}
\description{
This function provides estimates of  Body Cell Mass (BCM) using sex, H,W
}
