% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{BD}
\alias{BD}
\title{This function provides estimates of Body Density (BD) using  W, H, WC, HC, age, sex}
\usage{
BD(W, H, WC, HC, age, sex)
}
\arguments{
\item{W}{weight}

\item{H}{height}

\item{WC}{waist circumference}

\item{HC}{hip circumference}

\item{age}{age}

\item{sex}{sex}

\item{BD}{Body Density}
}
\description{
This function provides estimates of Body Density (BD) using  W, H, WC, HC, age, sex
}
