% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{BFM}
\alias{BFM}
\title{This function provides estimates of Body Fat Mass (BFM) using  W, TBF}
\usage{
BFM(W, TBF)
}
\arguments{
\item{W}{weight}

\item{TBF}{total body fat percentage}

\item{BFM}{Body Fat Mass}
}
\description{
This function provides estimates of Body Fat Mass (BFM) using  W, TBF
}
