% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{BFperc}
\alias{BFperc}
\title{This function provides estimates of bodyfat percentage  (BFperc) using age, sex, WC, H, W, HC}
\usage{
BFperc(sex, age, WC, H, W, HC)
}
\arguments{
\item{sex}{sex}

\item{age}{age}

\item{WC}{waist circumference}

\item{H}{height}

\item{W}{weight}

\item{HC}{hip circumference}

\item{BFperc}{body fat percentage}
}
\description{
This function provides estimates of bodyfat percentage  (BFperc) using age, sex, WC, H, W, HC
}
