% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{BMI}
\alias{BMI}
\title{This function provides estimates of body mass index (BMI) using W and H}
\usage{
BMI(W, H)
}
\arguments{
\item{W}{weight}

\item{H}{height}

\item{BMI}{body mass index}
}
\description{
This function provides estimates of body mass index (BMI) using W and H
}
