% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{BMI_l_SDS_V}
\alias{BMI_l_SDS_V}
\title{This function provides estimates of Body Mass Index-for-length Standard Deviation Score Velocity (BMI-L-SDS-V) using bmi_l_sds_t1, bmi_l_sds_t2, t1_age, t2_age}
\usage{
BMI_l_SDS_V(bmi_l_sds_t1, bmi_l_sds_t2, t1_age, t2_age)
}
\arguments{
\item{bmi_l_sds_t1}{bmi_l_sds_t1}

\item{bmi_l_sds_t2}{bmi_l_sds_t2}

\item{t1_age}{t1_age}

\item{t2_age}{t2_age}

\item{BMI_l_SDS_V}{Body Mass Index-for-length Standard Deviation Score Velocity}
}
\description{
This function provides estimates of Body Mass Index-for-length Standard Deviation Score Velocity (BMI-L-SDS-V) using bmi_l_sds_t1, bmi_l_sds_t2, t1_age, t2_age
}
