% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{BMIFORAGE}
\alias{BMIFORAGE}
\title{This function provides estimates of Body Mass Index for Age (BMIFORAGE)using  age,sex, BMI}
\usage{
BMIFORAGE(age, sex, BMI)
}
\arguments{
\item{age}{age}

\item{sex}{sex}

\item{BMI}{body mass index}

\item{BMIFORAGE}{Body Mass Index for Age}
}
\description{
This function provides estimates of Body Mass Index for Age (BMIFORAGE)using  age,sex, BMI
}
