% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{BMISDS}
\alias{BMISDS}
\title{This function provides estimates of BMI Standard Deviation Score (BMISDS) using  BMI, age,sex}
\usage{
BMISDS(BMI, sex)
}
\arguments{
\item{BMI}{body mass index}

\item{sex}{sex}

\item{BMISDS}{BMI Standard Deviation Score}
}
\description{
This function provides estimates of BMI Standard Deviation Score (BMISDS) using  BMI, age,sex
}
