% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{BMISDSV}
\alias{BMISDSV}
\title{This function provides estimates of Body Mass Index Standard Deviation Score Velocity (BMISDSV) using bmi_sds_t1, bmi_sds_t2, t1_age, t2_age}
\usage{
BMISDSV(bmi_sds_t1, bmi_sds_t2, t1_age, t2_age)
}
\arguments{
\item{bmi_sds_t1}{bmi_sds_t1}

\item{bmi_sds_t2}{bmi_sds_t2}

\item{t1_age}{t1_age}

\item{t2_age}{t2_age}

\item{BMISDSV}{Body Mass Index Standard Deviation Score Velocity}
}
\description{
This function provides estimates of Body Mass Index Standard Deviation Score Velocity (BMISDSV) using bmi_sds_t1, bmi_sds_t2, t1_age, t2_age
}
