% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{BMIV}
\alias{BMIV}
\title{This function provides estimates of Body Mass Index Velocity (BMIV) using weights, heights, times}
\usage{
BMIV(weights, heights, times)
}
\arguments{
\item{weights}{weights}

\item{heights}{heights}

\item{times}{times}

\item{BMIV}{Body Mass Index Velocity}
}
\description{
This function provides estimates of Body Mass Index Velocity (BMIV) using weights, heights, times
}
