% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{BMIZ}
\alias{BMIZ}
\title{This function provides estimates of Z-score for BMI (BMIZ) using  BMI, sex}
\usage{
BMIZ(BMI, sex)
}
\arguments{
\item{BMI}{body mass index}

\item{sex}{sex}

\item{BMIZ}{BMI Z-score}
}
\description{
This function provides estimates of Z-score for BMI (BMIZ) using  BMI, sex
}
