% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{bmiz_V}
\alias{bmiz_V}
\title{This function provides estimates of Body Mass Index-for-age Standard Deviation Score Velocity (BMIZ-V) using bmiz_t1, bmiz_t2, t1_age, t2_age}
\usage{
bmiz_V(bmiz_t1, bmiz_t2, t1_age, t2_age)
}
\arguments{
\item{bmiz_t1}{bmiz_t1}

\item{bmiz_t2}{bmiz_t2}

\item{t1_age}{t1_age}

\item{t2_age}{t2_age}

\item{bmiz_V}{Body Mass Index-for-age Standard Deviation Score Velocity}
}
\description{
This function provides estimates of Body Mass Index-for-age Standard Deviation Score Velocity (BMIZ-V) using bmiz_t1, bmiz_t2, t1_age, t2_age
}
