% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{BMR}
\alias{BMR}
\title{This function provides estimates of Basal Metabolic Rate (BMR) using W, H, age, sex}
\usage{
BMR(W, H, age, sex)
}
\arguments{
\item{W}{weight}

\item{H}{height}

\item{age}{age
#'@export}

\item{sex}{sex}

\item{BMR}{Basal Metabolic Rate}
}
\description{
This function provides estimates of Basal Metabolic Rate (BMR) using W, H, age, sex
}
