% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{BRI}
\alias{BRI}
\title{This function provides estimates of body roundness index (BRI) using  WC , H}
\usage{
BRI(WC, H)
}
\arguments{
\item{WC}{waist circumference}

\item{H}{height}

\item{BRI}{body roundness index}
}
\description{
This function provides estimates of body roundness index (BRI) using  WC , H
}
