% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{broca_index_BI}
\alias{broca_index_BI}
\title{This function provides estimates of broca index (BI) using  sex, W, H}
\usage{
broca_index_BI(H, sex)
}
\arguments{
\item{H}{height}

\item{sex}{sex}

\item{broca_index_BI}{broca index}
}
\description{
This function provides estimates of broca index (BI) using  sex, W, H
}
