% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{BSA}
\alias{BSA}
\title{This function provides estimates of body surface area (BSA) using  W, H}
\usage{
BSA(W, H)
}
\arguments{
\item{W}{weight}

\item{H}{height}

\item{BSA}{body surface area}
}
\description{
This function provides estimates of body surface area (BSA) using  W, H
}
