% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{BST}
\alias{BST}
\title{This function provides estimates of Biceps Skinfold Thickness (BST) using biceps_skinfold_measurement}
\usage{
BST(biceps_skinfold_measurement)
}
\arguments{
\item{biceps_skinfold_measurement}{biceps_skinfold_measurement}

\item{BST}{Biceps Skinfold Thickness}
}
\description{
This function provides estimates of Biceps Skinfold Thickness (BST) using biceps_skinfold_measurement
}
