% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{BVI}
\alias{BVI}
\title{This function provides estimates of Body Volume Index (BVI) using  W,WC, HC, H}
\usage{
BVI(W, H, WC, HC)
}
\arguments{
\item{W}{weight}

\item{H}{height}

\item{WC}{waist circumference}

\item{HC}{hip circumference}

\item{BVI}{Body Volume Index}
}
\description{
This function provides estimates of Body Volume Index (BVI) using  W,WC, HC, H
}
