% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{CI}
\alias{CI}
\title{This function provides estimates of conicity index (CI) using  WC, W, H}
\usage{
CI(WC, W, H)
}
\arguments{
\item{WC}{waist circumference}

\item{W}{weight}

\item{H}{height}

\item{CI}{conicity index}
}
\description{
This function provides estimates of conicity index (CI) using  WC, W, H
}
