% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{CRRorCRII}
\alias{CRRorCRII}
\title{This function provides estimates of Cardiovascular Risk Ratio (CRRorCRI I) using total_chol_mgdl, HDL_mgdl}
\usage{
CRRorCRII(total_chol_mgdl, HDL_mgdl)
}
\arguments{
\item{total_chol_mgdl, }{total cholesterol (mgdl)}

\item{HDL_mgdl}{high density i (mgdl)}

\item{CRRorCRII}{Cardiovascular Risk Ratio (CRR/CRI I)}
}
\description{
This function provides estimates of Cardiovascular Risk Ratio (CRRorCRI I) using total_chol_mgdl, HDL_mgdl
}
