% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{CST}
\alias{CST}
\title{This function provides estimates of Calf Skinfold Thickness (CST) using calf}
\usage{
CST(calf)
}
\arguments{
\item{calf}{calf}

\item{CST}{Calf Skinfold Thickness}
}
\description{
This function provides estimates of Calf Skinfold Thickness (CST) using calf
}
