% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{ECM}
\alias{ECM}
\title{This function provides estimates of Extracellular Mass (ECM) using BCM,W}
\usage{
ECM(W, BCM)
}
\arguments{
\item{W}{weight}

\item{BCM}{Body Cell Mass}

\item{ECM}{Extracellular Mass}
}
\description{
This function provides estimates of Extracellular Mass (ECM) using BCM,W
}
