% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{ECW}
\alias{ECW}
\title{This function provides estimates of Extracellular Water (ECW) using TBW,ICW}
\usage{
ECW(TBW, ICW)
}
\arguments{
\item{TBW}{Total Body Water}

\item{ICW}{Intracellular Water}

\item{ECW}{Extracellular Water}
}
\description{
This function provides estimates of Extracellular Water (ECW) using TBW,ICW
}
