% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{EER}
\alias{EER}
\title{This function provides estimates of Energy Requirement (EER) using bmr, tef, pa, pal}
\usage{
EER(bmr, tef, pa, pal)
}
\arguments{
\item{bmr}{Basal Metabolic Rate}

\item{tef}{Thermic Effect of Food}

\item{pa}{Physical Activity Level}

\item{pal}{sex
#'@export}

\item{EER}{Energy Requirement}
}
\description{
This function provides estimates of Energy Requirement (EER) using bmr, tef, pa, pal
}
