% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{FFBD}
\alias{FFBD}
\title{This function provides estimates of Fat-Free Body Density (FFBD) using  BD, TBF}
\usage{
FFBD(BD, TBF)
}
\arguments{
\item{BD}{Body Density}

\item{TBF}{total body fat percentage}
}
\description{
This function provides estimates of Fat-Free Body Density (FFBD) using  BD, TBF
}
