% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{FFM}
\alias{FFM}
\title{This function provides estimates of fat free mass (FFM) using  W,BFperc}
\usage{
FFM(W, BFperc)
}
\arguments{
\item{W}{weight}

\item{BFperc}{body fat percentage}

\item{FFM}{fat free mass}
}
\description{
This function provides estimates of fat free mass (FFM) using  W,BFperc
}
