% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{FFMI}
\alias{FFMI}
\title{This function provides estimates of fat free mass index (FFMI) using  FFM, H}
\usage{
FFMI(H, FFM)
}
\arguments{
\item{H}{height}

\item{FFM}{fat free mass}

\item{FFMI}{relative fat mass}
}
\description{
This function provides estimates of fat free mass index (FFMI) using  FFM, H
}
