% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{GV}
\alias{GV}
\title{This function provides estimates of Growth Velocity (GV) using H, times}
\usage{
GV(H, times)
}
\arguments{
\item{H}{height}

\item{times}{times}

\item{GV}{Growth Velocity}
}
\description{
This function provides estimates of Growth Velocity (GV) using H, times
}
