% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{HBE}
\alias{HBE}
\title{This function provides estimates of Harris-Benedict Equation (HBE)(is same as BMR) using  W,H,age,sex}
\usage{
HBE(W, H, age, sex)
}
\arguments{
\item{W}{weight}

\item{H}{height}

\item{age}{age}

\item{sex}{sex}

\item{HBE}{Harris-Benedict Equation (HBE)(is same as BMR)}
}
\description{
This function provides estimates of Harris-Benedict Equation (HBE)(is same as BMR) using  W,H,age,sex
}
