% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{HC}
\alias{HC}
\title{This function provides estimates of hip circumference (HC)}
\usage{
HC(HC)
}
\arguments{
\item{HC}{Hip Circumference}
}
\description{
This function provides estimates of hip circumference (HC)
}
