% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{hgs_to_bw_ratio}
\alias{hgs_to_bw_ratio}
\title{This function provides estimates of Handgrip Strength-to-Body Weight Ratio using  handgrip_strength, W}
\usage{
hgs_to_bw_ratio(handgrip_strength, W)
}
\arguments{
\item{handgrip_strength}{handgrip_strength}

\item{W}{weight}

\item{hgs_to_bw_ratio}{Handgrip Strength-to-Body Weight Ratio}
}
\description{
This function provides estimates of Handgrip Strength-to-Body Weight Ratio using  handgrip_strength, W
}
