% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{HI}
\alias{HI}
\title{This function provides estimates of hip-index(HI) using  HC, W, H}
\usage{
HI(W, H, HC)
}
\arguments{
\item{W}{weight}

\item{H}{height}

\item{HC}{hip circumference}

\item{HI}{hip index}
}
\description{
This function provides estimates of hip-index(HI) using  HC, W, H
}
