% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{HV}
\alias{HV}
\title{This function provides estimates of Height Velocity (HV) using heights, times}
\usage{
HV(heights, times)
}
\arguments{
\item{heights}{heights}

\item{times}{times}

\item{HV}{Height Velocity}
}
\description{
This function provides estimates of Height Velocity (HV) using heights, times
}
