% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{HWR}
\alias{HWR}
\title{This function provides estimates of hip-to-waist ratio (HWR) using  WC, HC}
\usage{
HWR(HC, WC)
}
\arguments{
\item{HC}{hip circumference}

\item{WC}{waist circumference}

\item{HWR}{hip-to-waist ratio}
}
\description{
This function provides estimates of hip-to-waist ratio (HWR) using  WC, HC
}
