% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{IBW}
\alias{IBW}
\title{This function provides estimates of Ideal Body Weight (IBW) using  H,sex}
\usage{
IBW(H, sex)
}
\arguments{
\item{H}{height}

\item{sex}{sex}

\item{IBW}{Ideal Body weight}
}
\description{
This function provides estimates of Ideal Body Weight (IBW) using  H,sex
}
