% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{ICW}
\alias{ICW}
\title{This function provides estimates of Intracellular Water (ICW) using TBW,ECM}
\usage{
ICW(TBW, ECM)
}
\arguments{
\item{TBW}{Total Body Water}

\item{ECM}{Extracellular Mass}

\item{ICW}{Intracellular Water}
}
\description{
This function provides estimates of Intracellular Water (ICW) using TBW,ECM
}
