% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{KMA}
\alias{KMA}
\title{This function provides estimates of Katch-McArdle Equation (KMA) using  LBM}
\usage{
KMA(LBM)
}
\arguments{
\item{LBM}{lean body mass}

\item{KMA}{Katch-McArdle Equation}
}
\description{
This function provides estimates of Katch-McArdle Equation (KMA) using  LBM
}
