% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{LAP}
\alias{LAP}
\title{This function provides estimates of Lipid Accumulation Product (LAP) by sex using sex, WC, triglycerides_mgdl}
\usage{
LAP(sex, WC, triglycerides_mgdl)
}
\arguments{
\item{sex}{sex}

\item{WC}{waist circumference}

\item{triglycerides_mgdl, }{triglycerides (mgdl)}

\item{LAP}{Lipid Accumulation Product by sex}
}
\description{
This function provides estimates of Lipid Accumulation Product (LAP) by sex using sex, WC, triglycerides_mgdl
}
