% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{LBM}
\alias{LBM}
\title{This function provides estimates of lean body mass (LBM) using  sex, W, H}
\usage{
LBM(W, H, sex)
}
\arguments{
\item{W}{weight}

\item{H}{height}

\item{sex}{sex}

\item{LBM}{relative fat mass}
}
\description{
This function provides estimates of lean body mass (LBM) using  sex, W, H
}
