% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{LBMI}
\alias{LBMI}
\title{This function provides estimates of Lean Body Mass Index (LBMI) using  LBM, H}
\usage{
LBMI(LBM, H)
}
\arguments{
\item{LBM}{Lean body mass}

\item{H}{height}

\item{LBMI}{Lean Body Mass Index}
}
\description{
This function provides estimates of Lean Body Mass Index (LBMI) using  LBM, H
}
