% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{LBMI_sex}
\alias{LBMI_sex}
\title{This function provides estimates of Lean Body Mass Index (LBMI_sex) by sex using LBM, H,sex}
\usage{
LBMI_sex(LBM, H, sex)
}
\arguments{
\item{LBM}{lean body mass}

\item{H}{height}

\item{sex}{sex}

\item{LBMI_sex}{Lean Body Mass Index by sex}
}
\description{
This function provides estimates of Lean Body Mass Index (LBMI_sex) by sex using LBM, H,sex
}
