% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{LDL}
\alias{LDL}
\title{This function provides estimates of  LDL using total_chol, HDL_chol, triglycerides}
\usage{
LDL(total_chol, HDL_chol, triglycerides)
}
\arguments{
\item{total_chol, }{total cholesterol (mgdl)}

\item{HDL_chol}{high density liproprotein chol (mgdl)}

\item{triglycerides}{triglycerides (mgdl)}

\item{LDL}{low-density lipoprotein cholesterol}
}
\description{
This function provides estimates of  LDL using total_chol, HDL_chol, triglycerides
}
