% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{LHRorCRIII}
\alias{LHRorCRIII}
\title{This function provides estimates of Cardiovascular Risk Index II (LHRorCRIII) or LDL-HDL Ratio (LHR) using total_chol_mgdl, LDL_mgdl, HDL_mgdl}
\usage{
LHRorCRIII(total_chol_mgdl, LDL_mgdl, HDL_mgdl)
}
\arguments{
\item{total_chol_mgdl, }{total cholesterol (mgdl)}

\item{LDL_mgdl}{low high density liproprotein (mgdl)}

\item{HDL_mgdl}{high density liproprotein (mgdl)}

\item{LHRorCRIII}{Cardiovascular Risk Index II (CRI II) or LDL-HDL Ratio (LHR)}
}
\description{
This function provides estimates of Cardiovascular Risk Index II (LHRorCRIII) or LDL-HDL Ratio (LHR) using total_chol_mgdl, LDL_mgdl, HDL_mgdl
}
