% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{MAMC}
\alias{MAMC}
\title{This function provides estimates of Mid-Arm Muscle Circumference (MAMC) using  midarm_circumference, triceps_skinfold_thickness}
\usage{
MAMC(midarm_circumference, triceps_skinfold_thickness)
}
\arguments{
\item{midarm_circumference}{midarm_circumference}

\item{triceps_skinfold_thickness}{triceps_skinfold_thickness}

\item{MAMC}{Handgrip Strength-to-Body Weight Ratio}
}
\description{
This function provides estimates of Mid-Arm Muscle Circumference (MAMC) using  midarm_circumference, triceps_skinfold_thickness
}
