% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{MPH}
\alias{MPH}
\title{This function provides estimates of ## Mid-Parental Height (MPH) using  mother_height, father_height, sex}
\usage{
MPH(mother_height, father_height, sex)
}
\arguments{
\item{mother_height}{mother height}

\item{father_height}{father height}

\item{sex}{sex}

\item{MPH}{## Mid-Parental Height}
}
\description{
This function provides estimates of ## Mid-Parental Height (MPH) using  mother_height, father_height, sex
}
