% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{MSJ}
\alias{MSJ}
\title{This function provides estimates of Mifflin-St Jeor Equation (MSJ) using  W,H,age,sex}
\usage{
MSJ(W, H, age, sex)
}
\arguments{
\item{W}{weight}

\item{H}{height}

\item{age}{age}

\item{sex}{sex}

\item{MSJ}{Mifflin-St Jeor Equation}
}
\description{
This function provides estimates of Mifflin-St Jeor Equation (MSJ) using  W,H,age,sex
}
