% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{PI_or_RI}
\alias{PI_or_RI}
\title{This function provides estimates of rohrer's or ponderal index (PI_or_RI) using  sex, W, H}
\usage{
PI_or_RI(H, W)
}
\arguments{
\item{H}{height}

\item{W}{weight}

\item{PI_or_RI}{rohrer's or ponderal index}
}
\description{
This function provides estimates of rohrer's or ponderal index (PI_or_RI) using  sex, W, H
}
