% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{RFM}
\alias{RFM}
\title{This function provides estimates of relative fat mass (RFM) using  sex, WC, H}
\usage{
RFM(WC, H, sex)
}
\arguments{
\item{WC}{waist circumference}

\item{H}{height}

\item{sex}{sex}

\item{RFM}{relative fat mass}
}
\description{
This function provides estimates of relative fat mass (RFM) using  sex, WC, H
}
