% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{RMR}
\alias{RMR}
\title{This function provides estimates of Resting Metabolic Rate (RMR) using BMR}
\usage{
RMR(BMR)
}
\arguments{
\item{BMR}{Basal Metabolic Rate
#'@export}

\item{RMR}{Resting Metabolic Rate}
}
\description{
This function provides estimates of Resting Metabolic Rate (RMR) using BMR
}
