% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{RPI}
\alias{RPI}
\title{This function provides estimates of reciprocal ponderal index (RPI) using  sex, W, H}
\usage{
RPI(H, W)
}
\arguments{
\item{H}{height}

\item{W}{weight}

\item{RPI}{rohrer's or ponderal index}
}
\description{
This function provides estimates of reciprocal ponderal index (RPI) using  sex, W, H
}
