% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{RW}
\alias{RW}
\title{This function provides estimates of Relative Weight (RW) using  W, H,sex}
\usage{
RW(W, H, sex)
}
\arguments{
\item{W}{weight}

\item{H}{height}

\item{sex}{sex}

\item{RW}{Relative Weight}
}
\description{
This function provides estimates of Relative Weight (RW) using  W, H,sex
}
