% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{SAD}
\alias{SAD}
\title{This function provides estimates of Sagittal Abdominal Diameter (SAD) using  WC, H}
\usage{
SAD(WC, H)
}
\arguments{
\item{WC}{waist circumference}

\item{H}{height}

\item{SAD}{Sagittal Abdominal Diameter}
}
\description{
This function provides estimates of Sagittal Abdominal Diameter (SAD) using  WC, H
}
