% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{sfd}
\alias{sfd}
\title{This function provides estimates of  Skinfold Density (sfd) using triceps, subscapular, suprailiac}
\usage{
sfd(triceps, subscapular, suprailiac)
}
\arguments{
\item{triceps}{triceps}

\item{subscapular}{subscapular}

\item{suprailiac}{suprailiac}

\item{sfd}{Skinfold Density}
}
\description{
This function provides estimates of  Skinfold Density (sfd) using triceps, subscapular, suprailiac
}
