% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{SFPerc}
\alias{SFPerc}
\title{This function provides estimates of Skinfold Fat Percentage (SFPerc) using sex, age, triceps, abdominal,subscapular, suprailiac}
\usage{
SFPerc(sex, triceps, age, subscapular, suprailiac, abdominal)
}
\arguments{
\item{sex}{sex}

\item{triceps}{triceps}

\item{age}{age}

\item{subscapular}{subscapular}

\item{suprailiac}{suprailiac}

\item{abdominal}{abdominal}

\item{SFPerc}{Skinfold Fat Percentage}
}
\description{
This function provides estimates of Skinfold Fat Percentage (SFPerc) using sex, age, triceps, abdominal,subscapular, suprailiac
}
