% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{SIT}
\alias{SIT}
\title{This function provides estimates of Suprailiac Skinfold Thickness (SIT) using suprailiac_skinfold_measurement}
\usage{
SIT(suprailiac_skinfold_measurement)
}
\arguments{
\item{suprailiac_skinfold_measurement}{suprailiac_skinfold_measurement}

\item{SIT}{Suprailiac Skinfold Thickness}
}
\description{
This function provides estimates of Suprailiac Skinfold Thickness (SIT) using suprailiac_skinfold_measurement
}
