% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{SST}
\alias{SST}
\title{This function provides estimates of Subscapular Skinfold Thickness (SST) using subscapular_skinfold_measurement}
\usage{
SST(subscapular_skinfold_measurement)
}
\arguments{
\item{subscapular_skinfold_measurement}{subscapular_skinfold_measurement}

\item{SST}{Subscapular Skinfold Thickness}
}
\description{
This function provides estimates of Subscapular Skinfold Thickness (SST) using subscapular_skinfold_measurement
}
