% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{SSTK}
\alias{SSTK}
\title{This function provides estimates of Sum of Skinfold Thicknesses (SSTK) using triceps, subscapular, suprailiac, thigh, calf}
\usage{
SSTK(triceps, subscapular, suprailiac, thigh, calf)
}
\arguments{
\item{triceps}{triceps}

\item{subscapular}{subscapular}

\item{suprailiac}{suprailiac}

\item{thigh}{thigh}

\item{calf}{calf}

\item{SSTK}{Sum of Skinfold Thicknesses}
}
\description{
This function provides estimates of Sum of Skinfold Thicknesses (SSTK) using triceps, subscapular, suprailiac, thigh, calf
}
