% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{TBF}
\alias{TBF}
\title{This function provides estimates of Total Body Fat Percentage (TBF) using  W, H, age, sex, WC, HC}
\usage{
TBF(W, H, age, sex, WC, HC)
}
\arguments{
\item{W}{weight}

\item{H}{height}

\item{age}{age}

\item{sex}{sex}

\item{WC}{waist circumference}

\item{HC}{hip circumference}

\item{TBF}{Total Body Fat Percentage}
}
\description{
This function provides estimates of Total Body Fat Percentage (TBF) using  W, H, age, sex, WC, HC
}
