% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{TBW}
\alias{TBW}
\title{This function provides estimates of Total Body Water (TBW) using TBW,W}
\usage{
TBW(W)
}
\arguments{
\item{W}{weight}

\item{TBW}{Total Body Water}
}
\description{
This function provides estimates of Total Body Water (TBW) using TBW,W
}
