% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{TH}
\alias{TH}
\title{This function provides estimates of Target Height (TH) using  sex, father_height, mother_height}
\usage{
TH(sex, father_height, mother_height)
}
\arguments{
\item{sex}{sex}

\item{father_height}{father height}

\item{mother_height}{mother height}

\item{TH}{Target Height}
}
\description{
This function provides estimates of Target Height (TH) using  sex, father_height, mother_height
}
