% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{THR}
\alias{THR}
\title{This function provides estimates of  Triglyceride HDL Ratio (THR) using triglycerides, HDL}
\usage{
THR(triglycerides, HDL)
}
\arguments{
\item{triglycerides}{triglycerides (mgdl)}

\item{HDL}{high density liproprotein (mgdl)}

\item{THR}{Triglyceride HDL Ratio}
}
\description{
This function provides estimates of  Triglyceride HDL Ratio (THR) using triglycerides, HDL
}
