% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{TST}
\alias{TST}
\title{This function provides estimates of Triceps Skinfold Thickness (TST) using triceps_skinfold_measurement}
\usage{
TST(triceps_skinfold_measurement)
}
\arguments{
\item{triceps_skinfold_measurement}{triceps_skinfold_measurement}

\item{TST}{Triceps Skinfold Thickness}
}
\description{
This function provides estimates of Triceps Skinfold Thickness (TST) using triceps_skinfold_measurement
}
