% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{TSTN}
\alias{TSTN}
\title{This function provides estimates of Thigh Skinfold Thickness (TSTN) using thigh}
\usage{
TSTN(thigh)
}
\arguments{
\item{thigh}{thigh}

\item{TSTN}{Thigh Skinfold Thickness}
}
\description{
This function provides estimates of Thigh Skinfold Thickness (TSTN) using thigh
}
