% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{TyG}
\alias{TyG}
\title{This function provides estimates of Triglycerides glucose index (TyG) using glucose, triglycerides}
\usage{
TyG(glucose, triglycerides)
}
\arguments{
\item{glucose}{glucose}

\item{triglycerides}{triglycerides (mgdl)}

\item{TyG}{Triglycerides Glucose Index}
}
\description{
This function provides estimates of Triglycerides glucose index (TyG) using glucose, triglycerides
}
