% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{VAI}
\alias{VAI}
\title{This function provides estimates of viseral adiposity index (VAI) using  sex, WC,BMI, hdl, triglycerides}
\usage{
VAI(sex, WC, BMI, triglycerides, hdl)
}
\arguments{
\item{sex}{sex}

\item{WC}{waist circumference}

\item{BMI}{body mass index}

\item{triglycerides}{triglycerides}

\item{hdl}{hdl}

\item{VAI}{visceral adiposity index}
}
\description{
This function provides estimates of viseral adiposity index (VAI) using  sex, WC,BMI, hdl, triglycerides
}
