% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{VLDL}
\alias{VLDL}
\title{This function provides estimates of Very Low density Lipoprotein (VLDL) using triglycerides}
\usage{
VLDL(triglycerides)
}
\arguments{
\item{triglycerides}{triglycerides (mgdl)}

\item{VLDL}{Very Low density Lipoprotein}
}
\description{
This function provides estimates of Very Low density Lipoprotein (VLDL) using triglycerides
}
