% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{WC}
\alias{WC}
\title{This function provides estimates of waist circumference (WC)}
\usage{
WC(WC)
}
\arguments{
\item{WC}{Waist Circumference}
}
\description{
This function provides estimates of waist circumference (WC)
}
