% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{WHR}
\alias{WHR}
\title{This function provides estimates of waist hip ratio (WHR) using WC, HC}
\usage{
WHR(WC, HC)
}
\arguments{
\item{WC}{waist circumference}

\item{HC}{hip circumference}

\item{WHR}{waist to hip ratio}
}
\description{
This function provides estimates of waist hip ratio (WHR) using WC, HC
}
