% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{WHT_5R}
\alias{WHT_5R}
\title{This function provides estimates of waist to height.5 ratio (WHT-5R) using  WC, H}
\usage{
WHT_5R(WC, H)
}
\arguments{
\item{WC}{waist circumference}

\item{H}{height}

\item{WHT_5R}{waist to height.5 ratio}
}
\description{
This function provides estimates of waist to height.5 ratio (WHT-5R) using  WC, H
}
