% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{WHtR}
\alias{WHtR}
\title{This function provides estimates of waist height ratio (WHTR) using  WC, H}
\usage{
WHtR(WC, H)
}
\arguments{
\item{WC}{waist circumference}

\item{H}{height}

\item{WHtR}{waist to height ratio}
}
\description{
This function provides estimates of waist height ratio (WHTR) using  WC, H
}
