% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioAnta.R
\name{WWI}
\alias{WWI}
\title{This function provides estimates of weighted-weight index (WWI) using WC, HC, H}
\usage{
WWI(WC, HC, H)
}
\arguments{
\item{WC}{waist circumference}

\item{HC}{hip circumference}

\item{H}{height}

\item{WWI}{weighted-weight index}
}
\description{
This function provides estimates of weighted-weight index (WWI) using WC, HC, H
}
